refactor dense rydberg hamiltonian in tests
